move tournamentrole/channel consts to use for frontend as well
