+++ DiscordBot/commands/tournaments/mappool/edit.ts
@@ -3,7 +3,6 @@ import { Command } from "../../index";
 import respond from "../../../functions/respond";
 import { extractParameter } from "../../../functions/parameterFunctions";
 import { securityChecks } from "../../../functions/tournamentFunctions/securityChecks";
-import { TournamentRoleType } from "../../../../Models/tournaments/tournamentRole";
 import getUser from "../../../../Server/functions/get/getUser";
 import commandUser from "../../../functions/commandUser";
 import { loginResponse } from "../../../functions/loginResponse";
@@ -17,6 +16,7 @@ import { User } from "../../../../Models/user";
 import getCustomThread from "../../../functions/tournamentFunctions/getCustomThread";
 import { discordClient } from "../../../../Server/discord";
 import { profanityFilterStrong } from "../../../../Interfaces/comment";
+import { TournamentRoleType } from "../../../../Interfaces/tournament";
 
 async function run (m: Message | ChatInputCommandInteraction) {
     if (m instanceof ChatInputCommandInteraction)
