+++ DiscordBot/commands/tournaments/matchup/stopAutoLobby.ts
@@ -2,8 +2,8 @@ import Axios from "axios";
 import { ChatInputCommandInteraction, Message, SlashCommandBuilder } from "discord.js";
 import { config } from "node-config-ts";
 import { Command } from "../..";
+import { TournamentRoleType } from "../../../../Interfaces/tournament";
 import { Matchup } from "../../../../Models/tournaments/matchup";
-import { TournamentRoleType } from "../../../../Models/tournaments/tournamentRole";
 import { extractParameter } from "../../../functions/parameterFunctions";
 import respond from "../../../functions/respond";
 import { securityChecks } from "../../../functions/tournamentFunctions/securityChecks";
