+++ DiscordBot/functions/tournamentFunctions/getStaff.ts
@@ -1,6 +1,7 @@
 import { ChatInputCommandInteraction, GuildMember, Message, User as DiscordUser } from "discord.js";
+import { TournamentRoleType } from "../../../Interfaces/tournament";
 import { Tournament } from "../../../Models/tournaments/tournament";
-import { TournamentRole, TournamentRoleType } from "../../../Models/tournaments/tournamentRole";
+import { TournamentRole } from "../../../Models/tournaments/tournamentRole";
 import { User } from "../../../Models/user";
 import { discordClient } from "../../../Server/discord";
 import getUser from "../../../Server/functions/get/getUser";
